Add build flavors "playStore", "openSource", and "openSourceDev"
Uses Gradle build flavors to distinguish between the productive
release version and the current development version, in order to allow
parallel installation of a productive version and a development version
on the same device.

Adjusts all themes.xml to not use qualified name to allow build of
flavor with different applicationId but same package name
(see http://tools.android.com/tech-docs/new-build-system/applicationid-vs-packagename
and http://stackoverflow.com/questions/28159281/android-gradle-applicationid-setting-causing-no-resource-found-in-themes-xml).

Also replaces hard-coded references to package name in XML files for
preferences.

Instrumentation tests are hard-coded against the "openSourceDev" flavor,
as this flavor is used for the test runner in the current configuration.