Bump Supported API Versions
Drop support for API 7 (Eclair) as it has less than .1% market share on the Play Store.
Extend support all the way up to API 22 (Lollipop_MR1).
Since Eclair support is dropped, the Mockito workaround for testing is deleted also.
Additionally, to make the project work with the latest Android Studio, the test manifest is deleted.