Add textVisiblePassword to inputType of key
Some very popular alternative keyboards have started to ignore the
"textNoSuggestions" hint and provide auto-completion nonetheless.
Adding the "textVisiblePassword" hint works around this in many cases.

This is important, because adding the key to a potentially backup-ed and
possibly cloud-synced auto-completion set of keyboard will result in the
key no longer being secret.